refactor: home 화면 레이어 선택 전 후 ui 별도 파일로 분리
